Move enum and `JsonElement` adapter classes to separate class files
Their implementation is complex enough to probably justify having them in
their own class files, to make maintaining them easier.
For the users this should not be noticeable since these classes are still
in the Gson internal package.
